Fix typing errors for TypedDocumentString
Using the package with a slighly different TypedDocumentString didn’t w
work, so make it more robust by accepting a wider type range
